Module filesystem functionalities in ModuleInput to ModuleFilesystem
