+++ Engine/Module/ModuleFileSystem.cpp
@@ -29,3 +29,38 @@ bool ModuleFileSystem::Copy(const char* source, const char* destination)
 {
 	return true;
 }
+
+ModuleFileSystem::FileType ModuleFileSystem::GetFileType(const char *file_path) const
+{
+	std::string file_extension = GetFileExtension(file_path);
+
+	if (
+		file_extension == "png"
+		|| file_extension == "PNG"
+		|| file_extension == "dds"
+		|| file_extension == "DDS"
+		)
+	{
+		return ModuleFileSystem::FileType::TEXTURE;
+	}
+	if (
+		file_extension == "fbx"
+		|| file_extension == "FBX"
+		)
+	{
+		return ModuleFileSystem::FileType::MODEL;
+	}
+
+	return ModuleFileSystem::FileType::UNKNOWN;
+}
+
+
+std::string ModuleFileSystem::GetFileExtension(const char *file_path) const
+{
+	std::string file_path_string = std::string(file_path);
+	std::size_t found = file_path_string.find_last_of(".");
+
+	std::string file_extension = file_path_string.substr(found + 1, file_path_string.length());
+
+	return file_extension;
+}
